backend completed

+++ src/main/java/com/example/yazlab2_1/Repository/ReferansRepo.java
@@ -3,5 +3,5 @@ package com.example.yazlab2_1.Repository;
 import com.example.yazlab2_1.Entities.Referans;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ReferansRepo extends MongoRepository<Referans, Integer> {
+public interface ReferansRepo extends MongoRepository<Referans, Long> {
 }
